inference mel masking fix.
