+++ Pattern_Generator.py
@@ -17,7 +17,7 @@ hp = Recursive_Parse(yaml.load(
 
 using_Extension = [x.upper() for x in ['.wav', '.m4a', '.flac']]
 regex_Checker = re.compile('[A-Z,.?!\'\-\s]+')
-top_DB_Dict = {'LJ': 60, 'BC2013': 60, 'VCTK': 15, 'VC1': 23, 'VC2': 23, 'Libri': 23, 'CMUA': 60}  # VC1 and Libri is from 'https://github.com/CorentinJ/Real-Time-Voice-Cloning'
+top_DB_Dict = {'LJ': 60, 'BC2013': 60, 'VCTK': 15, 'VC1': 23, 'VC1T': 23, 'VC2': 23, 'Libri': 23, 'CMUA': 60}  # VC1 and Libri is from 'https://github.com/CorentinJ/Real-Time-Voice-Cloning'
 
 def Text_Filtering(text):
     remove_Letter_List = ['(', ')', '\"', '[', ']', ':', ';']
@@ -71,6 +71,16 @@ def Pattern_Generate(path, top_db= 60):
 def Pattern_File_Generate(path, speaker_ID, speaker, dataset, text= None, tag='', eval= False):
     pattern_Path = hp.Train.Eval_Pattern.Path if eval else hp.Train.Train_Pattern.Path
 
+    file = '{}.{}{}.PICKLE'.format(
+        speaker if dataset in speaker else '{}.{}'.format(dataset, speaker),
+        '{}.'.format(tag) if tag != '' else '',
+        os.path.splitext(os.path.basename(path))[0]
+        ).upper()
+    file = os.path.join(pattern_Path, dataset, speaker, file).replace("\\", "/")
+
+    if os.path.exists(file):
+        return
+
     try:
         audio, mel, pitch = Pattern_Generate(path, top_DB_Dict[dataset])
         assert mel.shape[0] == pitch.shape[0], 'Mel_shape != Pitch_shape {} != {}'.format(mel.shape, pitch.shape)
@@ -88,13 +98,8 @@ def Pattern_File_Generate(path, speaker_ID, speaker, dataset, text= None, tag=''
         print('Error: {} in {}'.format(e, path))
         return
 
-    file = '{}.{}{}.PICKLE'.format(
-        speaker if dataset in speaker else '{}.{}'.format(dataset, speaker),
-        '{}.'.format(tag) if tag != '' else '',
-        os.path.splitext(os.path.basename(path))[0]
-        ).upper()
 
-    os.makedirs(os.path.join(pattern_Path, dataset).replace('\\', '/'), exist_ok= True)
+    os.makedirs(os.path.join(pattern_Path, dataset, speaker).replace('\\', '/'), exist_ok= True)
     with open(os.path.join(pattern_Path, dataset, file).replace("\\", "/"), 'wb') as f:
         pickle.dump(new_Pattern_Dict, f, protocol=4)
 
@@ -246,15 +251,90 @@ def Libri_Info_Load(path, use_text= False):
     print('Libri info generated: {}'.format(len(paths)))
     return paths, text_Dict, speaker_Dict
 
+
+def VC1_Info_Load(path, use_text= False):
+    if use_text:
+        raise ValueError('VC1 does not support the text.')
+
+    paths = []
+    for root, _, files in os.walk(path):
+        for file in files:
+            file = os.path.join(root, file).replace('\\', '/')
+            if not os.path.splitext(file)[1].upper() in using_Extension:
+                continue
+            paths.append(file)
+    
+    speaker_Dict = {
+        path: 'VC1.{}'.format(path.split('/')[-3].upper())
+        for path in paths
+        }
+    tag_Dict = {
+        path: path.split('/')[-2].upper()
+        for path in paths
+        }
+
+    print('VC1 info generated: {}'.format(len(paths)))
+    return paths, speaker_Dict, tag_Dict
+
+def VC2_Info_Load(path, use_text= False):
+    if use_text:
+        raise ValueError('VC2 does not support the text.')
+
+    paths = []
+    for root, _, files in os.walk(path):
+        for file in files:
+            wav_File_Path = os.path.join(root, file).replace('\\', '/')
+            if not os.path.splitext(wav_File_Path)[1].upper() in using_Extension:
+                continue
+            paths.append(wav_File_Path)
+    
+    speaker_Dict = {
+        path: 'VC2.{}'.format(path.split('/')[-3].upper())
+        for path in paths
+        }
+    tag_Dict = {
+        path: path.split('/')[-2].upper()
+        for path in paths
+        }
+
+    print('VC2 info generated: {}'.format(len(paths)))
+    return paths, speaker_Dict, tag_Dict
+
+
+def VC1T_Info_Load(path, use_text= False):
+    if use_text:
+        raise ValueError('VC1-Test does not support the text.')
+
+    paths = []
+    for root, _, files in os.walk(path):
+        for file in files:
+            file = os.path.join(root, file).replace('\\', '/')
+            if not os.path.splitext(file)[1].upper() in using_Extension:
+                continue
+            paths.append(file)
+    
+    speaker_Dict = {
+        path: 'VC1T.{}'.format(path.split('/')[-3].upper())
+        for path in paths
+        }
+    tag_Dict = {
+        path: path.split('/')[-2].upper()
+        for path in paths
+        }
+
+    print('VC1T info generated: {}'.format(len(paths)))
+    return paths, speaker_Dict, tag_Dict
+
+
 def Speaker_Index_Dict_Generate(speaker_Dict):
     return {
         speaker: index
         for index, speaker in enumerate(sorted(set(speaker_Dict.values())))
         }
 
-def Split_Eval(paths, eval_ratio= 0.001):
+def Split_Eval(paths, eval_ratio= 0.001, min_Eval= 1):
     shuffle(paths)
-    index = int(len(paths) * eval_ratio)
+    index = max(int(len(paths) * eval_ratio), min_Eval)
     return paths[index:], paths[:index]
 
 def Metadata_Generate(eval= False, use_text= False):
@@ -271,20 +351,26 @@ def Metadata_Generate(eval= False, use_text= False):
         'File_List': [],
         'Audio_Length_Dict': {},
         'Mel_Length_Dict': {},
-        'Pitch_Length_Dict': {},        
+        'Pitch_Length_Dict': {},
         'Speaker_ID_Dict': {},
         'Speaker_Dict': {},
         'Dataset_Dict': {},
+        'File_List_by_Speaker_Dict': {},
         }
     if use_text:
         new_Metadata_Dict['Text_Length_Dict'] = {}
+
+    files_TQDM = tqdm(
+        total= sum([len(files) for root, _, files in os.walk(pattern_Path)]),
+        desc= 'Eval_Pattern' if eval else 'Train_Pattern'
+        )
 
     for root, _, files in os.walk(pattern_Path):
         for file in files:
             with open(os.path.join(root, file).replace("\\", "/"), "rb") as f:
                 pattern_Dict = pickle.load(f)
 
-            file = os.path.join(os.path.basename(root), file).replace("\\", "/")
+            file = os.path.join(root, file).replace("\\", "/").replace(pattern_Path, '').lstrip('/')
             try:
                 if not all([
                     key in ('Audio', 'Mel', 'Pitch', 'Speaker_ID', 'Speaker', 'Dataset', 'Text' if use_text else '')
@@ -293,15 +379,19 @@ def Metadata_Generate(eval= False, use_text= False):
                     continue
                 new_Metadata_Dict['Audio_Length_Dict'][file] = pattern_Dict['Audio'].shape[0]
                 new_Metadata_Dict['Mel_Length_Dict'][file] = pattern_Dict['Mel'].shape[0]
-                new_Metadata_Dict['Pitch_Length_Dict'][file] = pattern_Dict['Pitch'].shape[0]                
+                new_Metadata_Dict['Pitch_Length_Dict'][file] = pattern_Dict['Pitch'].shape[0]
                 new_Metadata_Dict['Speaker_ID_Dict'][file] = pattern_Dict['Speaker_ID']
                 new_Metadata_Dict['Speaker_Dict'][file] = pattern_Dict['Speaker']
                 new_Metadata_Dict['Dataset_Dict'][file] = pattern_Dict['Dataset']
                 new_Metadata_Dict['File_List'].append(file)
+                if not pattern_Dict['Speaker'] in new_Metadata_Dict['File_List_by_Speaker_Dict'].keys():
+                    new_Metadata_Dict['File_List_by_Speaker_Dict'][pattern_Dict['Speaker']] = []
+                new_Metadata_Dict['File_List_by_Speaker_Dict'][pattern_Dict['Speaker']].append(file)
                 if use_text:
                     new_Metadata_Dict['Text_Length_Dict'][file] = len(pattern_Dict['Text'])
             except:
                 print('File \'{}\' is not correct pattern file. This file is ignored.'.format(file))
+            files_TQDM.update(1)
 
     with open(os.path.join(pattern_Path, metadata_File.upper()).replace("\\", "/"), 'wb') as f:
         pickle.dump(new_Metadata_Dict, f, protocol= 4)
@@ -319,13 +409,6 @@ def Token_Dict_Generate(text_Dict):
         {token: index for index, token in enumerate(['<S>', '<E>'] + sorted(tokens))},
         open(hp.Token_Path, 'w')
         )
-    
-    # #I don't use yaml.dump in this case to sort clearly.
-    # os.makedirs(os.path.dirname(hp.Token_Path), exist_ok= True)
-    # open(hp.Token_Path, 'w').write('\n'.join([
-    #     '\'{}\': {}'.format(token, index)
-    #     for index, token in enumerate(['<S>', '<E>'] + sorted(tokens))
-    #     ]))
 
 if __name__ == '__main__':
     argParser = argparse.ArgumentParser()
@@ -334,9 +417,14 @@ if __name__ == '__main__':
     argParser.add_argument("-cmua", "--cmua_path", required=False)
     argParser.add_argument("-vctk", "--vctk_path", required=False)
     argParser.add_argument("-libri", "--libri_path", required=False)
+    argParser.add_argument("-vc1", "--vc1_path", required=False)
+    argParser.add_argument("-vc2", "--vc2_path", required=False)
+
+    argParser.add_argument("-vc1t", "--vc1_test_path", required=False)
     
     argParser.add_argument("-text", "--use_text", action= 'store_true')
-    argParser.add_argument("-eval", "--eval_ratio", default= 0.001, type= float)
+    argParser.add_argument("-evalr", "--eval_ratio", default= 0.001, type= float)
+    argParser.add_argument("-evalm", "--eval_min", default= 1, type= int)
     argParser.add_argument("-mw", "--max_worker", default= 10, required=False, type= int)
 
     args = argParser.parse_args()
@@ -382,6 +470,25 @@ if __name__ == '__main__':
         speaker_Dict.update(libri_Speaker_Dict)
         dataset_Dict.update({path: 'Libri' for path in libri_Paths})
         tag_Dict.update({path: '' for path in libri_Paths})
+    if not args.vc1_path is None:
+        vc1_Paths, vc1_Speaker_Dict, vc1_Tag_Dict = VC1_Info_Load(path= args.vc1_path, use_text= args.use_text)
+        paths.extend(vc1_Paths)
+        speaker_Dict.update(vc1_Speaker_Dict)
+        dataset_Dict.update({path: 'VC1' for path in vc1_Paths})
+        tag_Dict.update(vc1_Tag_Dict)
+    if not args.vc2_path is None:
+        vc2_Paths, vc2_Speaker_Dict, vc2_Tag_Dict = VC2_Info_Load(path= args.vc2_path, use_text= args.use_text)
+        paths.extend(vc2_Paths)
+        speaker_Dict.update(vc2_Speaker_Dict)
+        dataset_Dict.update({path: 'VC2' for path in vc2_Paths})
+        tag_Dict.update(vc2_Tag_Dict)
+
+    if not args.vc1_test_path is None:
+        vc1t_Paths, vc1t_Speaker_Dict, vc1t_Tag_Dict = VC1T_Info_Load(path= args.vc1_test_path, use_text= args.use_text)
+        paths.extend(vc1t_Paths)
+        speaker_Dict.update(vc1t_Speaker_Dict)
+        dataset_Dict.update({path: 'VC1T' for path in vc1t_Paths})
+        tag_Dict.update(vc1t_Tag_Dict)
 
     if len(paths) == 0:
         raise ValueError('Total info count must be bigger than 0.')
@@ -392,7 +499,7 @@ if __name__ == '__main__':
     speaker_Index_Dict = Speaker_Index_Dict_Generate(speaker_Dict)
 
     train_Paths, eval_Paths = Split_Eval(paths, args.eval_ratio)
-    
+
     with PE(max_workers = args.max_worker) as pe:
         for _ in tqdm(            
             pe.map(
@@ -437,8 +544,8 @@ if __name__ == '__main__':
     Metadata_Generate(eval= True, use_text= args.use_text)
 
 
-
 # python Pattern_Generator.py -lj "D:\Pattern\ENG\LJSpeech" -bc2013 "D:\Pattern\ENG\BC2013" -cmua "D:\Pattern\ENG\CMUA" -vctk "D:\Pattern\ENG\VCTK" -libri "D:\Pattern\ENG\LibriTTS"
 # python Pattern_Generator.py -lj "D:\Pattern\ENG\LJSpeech" -vctk "D:\Pattern\ENG\VCTK" -libri "D:\Pattern\ENG\LibriTTS" -text
 # python Pattern_Generator.py -lj "D:\Pattern\ENG\LJSpeech" -text
 # python Pattern_Generator.py -lj /home/heejo/data/Eng/LJSpeech-1.1 -text
+# python Pattern_Generator.py -vc2 "D:\Pattern\ENG\VC2" -mw 1
